move dist file creation into a util
